src/test/kotlin/io/haerong22/ticketing/application/user/command/UserCommandTest.kt
package io.haerong22.ticketing.application.user.command

import io.haerong22.ticketing.domain.common.BadRequestException
import org.assertj.core.api.Assertions.assertThatThrownBy
import org.junit.jupiter.api.Test

class UserCommandTest {

    @Test
    fun `포인트 조회 시 userId는 음수로 초기화 하면 BadRequestException 이 발생한다`() {
        // given
        val userId = -1L

        // when, then
        assertThatThrownBy { UserCommand.GetPoint(userId) }
            .isInstanceOf(BadRequestException::class.java)
            .hasMessage("userId는 양수 값 입니다.")
    }

    @Test
    fun `포인트 충전 시 userId는 음수로 초기화 하면 BadRequestException 이 발생한다`() {
        // given
        val userId = -1L
        val amount = 0

        // when, then
        assertThatThrownBy { UserCommand.ChargePoint(userId, amount) }
            .isInstanceOf(BadRequestException::class.java)
            .hasMessage("userId는 양수 값 입니다.")
    }

    @Test
    fun `포인트 충전 시 amount는 음수로 초기화 하면 BadRequestException 이 발생한다`() {
        // given
        val userId = 1L
        val amount = -1

        // when, then
        assertThatThrownBy { UserCommand.ChargePoint(userId, amount) }
            .isInstanceOf(BadRequestException::class.java)
            .hasMessage("amount는 양수 값 입니다.")
    }
}
